handlers/extract/requirements.txt
runpod>=1.6.0
httpx>=0.26.0
requests>=2.31.0
pillow>=10.0.0
